chore(@clayui/forms): Move input-group to own section in storybook
