+++ packages/clay-form/stories/Input.stories.tsx
@@ -86,113 +86,3 @@ Textarea.args = {
 	readOnly: false,
 	sizing: undefined,
 };
-
-export const GroupSeparated = () => (
-	<div className="sheet">
-		<ClayForm.Group>
-			<ClayInput.Group>
-				<ClayInput.GroupItem shrink>
-					<ClayInput.GroupText>@</ClayInput.GroupText>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem>
-					<ClayInput
-						aria-label="Username"
-						placeholder="Username"
-						type="text"
-					/>
-				</ClayInput.GroupItem>
-			</ClayInput.Group>
-		</ClayForm.Group>
-		<ClayForm.Group>
-			<ClayInput.Group>
-				<ClayInput.GroupItem>
-					<ClayInput
-						aria-label="Email"
-						placeholder="Email"
-						type="text"
-					/>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem shrink>
-					<ClayInput.GroupText>@</ClayInput.GroupText>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem>
-					<ClayInput
-						aria-label="Email Host"
-						placeholder="liferay"
-						type="text"
-					/>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem shrink>
-					<ClayInput.GroupText>.com</ClayInput.GroupText>
-				</ClayInput.GroupItem>
-			</ClayInput.Group>
-		</ClayForm.Group>
-	</div>
-);
-
-export const GroupConnected = () => (
-	<div className="sheet">
-		<ClayForm.Group>
-			<ClayInput.Group>
-				<ClayInput.GroupItem prepend shrink>
-					<ClayInput.GroupText>@</ClayInput.GroupText>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem append>
-					<ClayInput
-						aria-label="Username"
-						placeholder="Username"
-						type="text"
-					/>
-				</ClayInput.GroupItem>
-			</ClayInput.Group>
-		</ClayForm.Group>
-		<ClayForm.Group>
-			<ClayInput.Group>
-				<ClayInput.GroupItem prepend>
-					<ClayInput
-						aria-label="Email"
-						placeholder="Email"
-						type="text"
-					/>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem prepend shrink>
-					<ClayInput.GroupText>@</ClayInput.GroupText>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem prepend>
-					<ClayInput
-						aria-label="Email Host"
-						placeholder="liferay"
-						type="text"
-					/>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem append shrink>
-					<ClayInput.GroupText>.com</ClayInput.GroupText>
-				</ClayInput.GroupItem>
-			</ClayInput.Group>
-		</ClayForm.Group>
-	</div>
-);
-
-export const GroupMixed = () => (
-	<div className="sheet">
-		<ClayForm.Group>
-			<ClayInput.Group>
-				<ClayInput.GroupItem shrink>
-					<ClayInput.GroupText>@</ClayInput.GroupText>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem prepend>
-					<ClayInput
-						aria-label="Username"
-						placeholder="Username"
-						type="text"
-					/>
-				</ClayInput.GroupItem>
-				<ClayInput.GroupItem append shrink>
-					<button className="btn btn-secondary" type="submit">
-						Submit
-					</button>
-				</ClayInput.GroupItem>
-			</ClayInput.Group>
-		</ClayForm.Group>
-	</div>
-);
